script now copies dist/k1ussnsst to .local/bin

+++ install_icon/install.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-cp ../dist/k1usnsst ~/.local/bin/
+if [ -f "../dist/k1usnsst" ]; then
+	cp ../dist/k1usnsst ~/.local/bin/
+fi
 
 xdg-icon-resource install --size 64 --context apps --mode user ../icon/K1USN-SST.png k6gte-k1usnsst
